Написал логику для сервиса объявлений, написал роли для каждого метода, изменил логику безопасности

+++ src/main/java/ru/skypro/homework/repository/AdRepository.java
@@ -3,5 +3,5 @@ package ru.skypro.homework.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.skypro.homework.model.AdModel;
 
-public interface AdRepository extends JpaRepository<AdModel, Long> {
+public interface AdRepository extends JpaRepository<AdModel, Integer> {
 }
