+++ src/main/java/ru/skypro/homework/repository/CommentRepository.java
@@ -3,5 +3,5 @@ package ru.skypro.homework.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.skypro.homework.model.CommentModel;
 
-public interface CommentRepository extends JpaRepository<CommentModel, Long> {
+public interface CommentRepository extends JpaRepository<CommentModel, Integer> {
 }
